Fix buffer overflow, signed integer overflow, and division by zero
This commit addresses three critical issues found during a deep code scan:
1.  **Buffer Overflow in `ItemDatabase::loadFromOtbGeneric`**: Fixed a buffer overflow where `ITEM_ATTR_WEIGHT` (8 bytes) was being read directly into `ItemType::weight` (4 bytes, float), overwriting adjacent memory (likely `attack` field). The fix involves reading into a temporary `double` and casting to `float`.
2.  **Signed Integer Overflow in `Outfit::getColorHash`**: Fixed undefined behavior caused by left-shifting signed integers into the sign bit. The operands are now cast to `uint32_t` before shifting.
3.  **Division by Zero in `IOMapOTBM::readMapNodes`**: Fixed a potential division by zero crash when updating the load progress bar for empty map files. Added a check for `f.size() > 0`.

These changes improve the stability and safety of the application.

+++ source/game/outfit.h
@@ -38,11 +38,11 @@ struct Outfit {
 	int lookMountFeet;
 
 	uint32_t getColorHash() const {
-		return lookHead << 24 | lookBody << 16 | lookLegs << 8 | lookFeet;
+		return static_cast<uint32_t>(lookHead) << 24 | static_cast<uint32_t>(lookBody) << 16 | static_cast<uint32_t>(lookLegs) << 8 | static_cast<uint32_t>(lookFeet);
 	}
 
 	uint32_t getMountColorHash() const {
-		return lookMountHead << 24 | lookMountBody << 16 | lookMountLegs << 8 | lookMountFeet;
+		return static_cast<uint32_t>(lookMountHead) << 24 | static_cast<uint32_t>(lookMountBody) << 16 | static_cast<uint32_t>(lookMountLegs) << 8 | static_cast<uint32_t>(lookMountFeet);
 	}
 };
